Only determine execute permissions on files we copy/move. (#3437)
* Permissions refactor.

* Add exec bit to all jobstores.

* Update fileJobStore.py

* Update abstractJobStore.py

* Have atomic_copy remember permissions.

* Restore comments.

* Update for apt packages first.
