+++ src/toil/jobStores/abstractJobStore.py
@@ -32,7 +32,7 @@ from toil.job import (CheckpointJobDescription,
                       JobException,
                       ServiceJobDescription)
 from toil.lib.memoize import memoize
-from toil.lib.misc import WriteWatchingStream
+from toil.lib.io import WriteWatchingStream
 from toil.lib.retry import ErrorCondition, retry
 
 logger = logging.getLogger(__name__)
@@ -955,7 +955,9 @@ class AbstractJobStore(ABC):
         appear in the local file system until the copy has completed.
 
         The file at the given local path may not be modified after this method returns!
-
+        
+        Note!  Implementations of readFile need to respect/provide the executable attribute on FileIDs.
+        
         :param str jobStoreFileID: ID of the file to be copied
 
         :param str localFilePath: the local path indicating where to place the contents of the
